fix state event emission, move tests (#105)
