+++ pymmcore_plus/_tests/test_core.py
@@ -14,7 +14,6 @@ from qtpy.QtCore import QObject
 from qtpy.QtCore import SignalInstance as QSignalInstance
 from useq import MDASequence
 
-import pymmcore_plus
 from pymmcore_plus import (
     CMMCorePlus,
     Configuration,
@@ -24,23 +23,6 @@ from pymmcore_plus import (
     PropertyType,
 )
 from pymmcore_plus.core.events import CMMCoreSignaler
-
-
-@pytest.fixture(params=["QSignal", "psygnal"], scope="function")
-def core(request):
-    core = CMMCorePlus()
-    if request.param == "psygnal":
-        core.events = CMMCoreSignaler()
-        core._callback_relay = pymmcore_plus.core._mmcore_plus.MMCallbackRelay(
-            core.events
-        )
-        core.registerCallback(core._callback_relay)
-    if not core.getDeviceAdapterSearchPaths():
-        pytest.fail(
-            "To run tests, please install MM with `python -m pymmcore_plus.install`"
-        )
-    core.loadSystemConfiguration()
-    return core
 
 
 def test_core(core: CMMCorePlus):
@@ -331,49 +313,6 @@ def test_property_schema(core: CMMCorePlus):
     assert isinstance(schema, dict)
     assert schema["title"] == "DCam"
     assert schema["properties"]["AllowMultiROI"] == {"type": "boolean"}
-
-
-def test_set_property_events(core: CMMCorePlus):
-    """Test that using setProperty always emits a propertyChanged event."""
-    mock = MagicMock()
-    core.events.propertyChanged.connect(mock)
-    core.setProperty("Camera", "Binning", "2")
-    mock.assert_called_once_with("Camera", "Binning", "2")
-
-    mock.reset_mock()
-    core.setProperty("Camera", "Binning", "1")
-    mock.assert_called_once_with("Camera", "Binning", "1")
-
-    mock.reset_mock()
-    core.setProperty("Camera", "Binning", "1")
-    mock.assert_not_called()  # value didn't change
-
-    # this is not a property that the DemoCamera emits...
-    # so with regular pymmcore, this would not be emitted.
-    core.setProperty("Camera", "AllowMultiROI", "1")
-    mock.assert_called_once_with("Camera", "AllowMultiROI", "1")
-
-
-def test_set_state_events(core: CMMCorePlus):
-    mock = MagicMock()
-    core.events.propertyChanged.connect(mock)
-    assert core.getState("Objective") == 1
-    core.setState("Objective", 3)
-    mock.assert_has_calls(
-        [
-            call("Objective", "State", "3"),
-            call("Objective", "Label", "Nikon 20X Plan Fluor ELWD"),
-        ]
-    )
-    assert core.getState("Objective") == 3
-
-    mock.reset_mock()
-    assert core.getState("Dichroic") == 0
-    core.setStateLabel("Dichroic", "Q505LP")
-    mock.assert_has_calls(
-        [call("Dichroic", "State", "1"), call("Dichroic", "Label", "Q505LP")]
-    )
-    assert core.getState("Dichroic") == 1
 
 
 def test_get_objectives(core: CMMCorePlus):
